+++ src/main/java/com/byplace/web/admin/user/AdminUserEdit.java
@@ -32,7 +32,7 @@ public class AdminUserEdit extends HttpServlet {
 		PrintWriter writer = response.getWriter();
 		HttpSession session = request.getSession();
 //		if(session.getAttribute("USER") != null && ((UserDTO)session.getAttribute("USER")).getUser_type().equals("관리자")) {
-			int user_no = Integer.parseInt(request.getParameter("user_no"));
+			long user_no = Long.parseLong(request.getParameter("user_no"));
 			AdminUserDAO adminUserDAO = new AdminUserDAO();
 			UserDTO dto = new UserDTO();
 			dto.setUser_no(user_no);
